requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-telegram-bot
python-multipart
jinja2
PyJWT
